refactor: move generate seed users to scriptUtils
Refactor the functionality that generate an array of seed users data
into scriptUtils file.
